Tweak push actions git diff statement
Looking for optimal way to check which files changed.
It appears, that this check is triggered after push, so need to look
at previous commit for diff.  Is this going to work on pull_request?
Probably want PR to run lint on all files changed in the etire PR branch
from first commit.